+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.11] 2021-03-17
 ### Added
 - Deployment of API server to Heroku
+- Add CLI option for registering services
 ### Changed
 - Server jar no longer deployed to github Maven repo